refactor: 함수컨벤션 - function선언 => 화살표함수 변경
